requirements.txt
Django==1.5.5
Markdown==2.3.1
django-filter==0.7
djangorestframework==2.3.8
wsgiref==0.1.2
